+++ src/agents/workflows/twitter/tools.ts
@@ -1,10 +1,10 @@
 import { TwitterApi } from '../../../services/twitter/types.js';
-import { createFetchMentionsTool } from '../../tools/fetchMentionsTool.js';
-import { createFetchTimelineTool } from '../../tools/fetchTimelineTool.js';
-import { createPostTweetTool } from '../../tools/postTweetTool.js';
+import { createFetchMentionsTool } from '../../tools/twitter/fetchMentionsTool.js';
+import { createFetchTimelineTool } from '../../tools/twitter/fetchTimelineTool.js';
+import { createPostTweetTool } from '../../tools/twitter/postTweetTool.js';
 import { createUploadToDsnTool } from '../../tools/uploadToDsnTool.js';
-import { createFetchMyRecentTweetsTool } from '../../tools/fetchMyRecentTweetsTool.js';
-import { createFetchMyRecentRepliesTool } from '../../tools/fetchMyRecentRepliesTool.js';
+import { createFetchMyRecentTweetsTool } from '../../tools/twitter/fetchMyRecentTweetsTool.js';
+import { createFetchMyRecentRepliesTool } from '../../tools/twitter/fetchMyRecentRepliesTool.js';
 export const createTools = (twitterApi: TwitterApi) => {
   const fetchTimelineTool = createFetchTimelineTool(twitterApi);
   const fetchMentionsTool = createFetchMentionsTool(twitterApi);
